Refactor season winner role updater
