+++ src/discord/seasonWinnerRoleUpdater/service/getTopUserIdsFromSortedMessageCounts.ts
@@ -1,9 +1,10 @@
-import type { UserMessageCount } from "src/discord/utils/messageCounts";
+import type { UserMessageCount } from "@/src/discord/stats/messageCounts";
 
 /**
  * Returns all user ids tied for the highest message count.
  */
-export const resolveSeasonWinnerUserIds = (
+export const getTopUserIdsFromSortedMessageCounts = (
+  // Message counts by user in descending order
   messageCounts: UserMessageCount[],
 ): string[] => {
   const topMessageCount = messageCounts[0]?.messageCount;
