+++ src/discord/seasonWinnerRoleUpdater/service/resolveSeasonWinnerUserIds.test.ts
@@ -1,9 +1,9 @@
-import { resolveSeasonWinnerUserIds } from "./resolveSeasonWinnerUserIds";
+import { getTopUserIdsFromSortedMessageCounts } from "./getTopUserIdsFromSortedMessageCounts";
 
-describe("resolveSeasonWinnerUserIds", () => {
+describe("getTopUserIdsFromSortedMessageCounts", () => {
   it("returns a unique winner", () => {
     expect(
-      resolveSeasonWinnerUserIds([
+      getTopUserIdsFromSortedMessageCounts([
         { userId: "user-1", messageCount: 5 },
         { userId: "user-2", messageCount: 3 },
       ]),
@@ -12,7 +12,7 @@ describe("resolveSeasonWinnerUserIds", () => {
 
   it("returns all tied winners", () => {
     expect(
-      resolveSeasonWinnerUserIds([
+      getTopUserIdsFromSortedMessageCounts([
         { userId: "user-1", messageCount: 5 },
         { userId: "user-2", messageCount: 5 },
         { userId: "user-3", messageCount: 1 },
@@ -21,6 +21,6 @@ describe("resolveSeasonWinnerUserIds", () => {
   });
 
   it("returns an empty list when no messages exist", () => {
-    expect(resolveSeasonWinnerUserIds([])).toEqual([]);
+    expect(getTopUserIdsFromSortedMessageCounts([])).toEqual([]);
   });
 });
